### 2024-12-03 #7558 @editorialbot
[PRE REVIEW]: LangFair: A Python Package for Assessing Bias and Fairness in Large Language Model Use Cases
**Submitting author:** @DylanBouchard (<a href="http://orcid.org/0009-0004-9233-2324">Dylan Bouchard</a>)
**Repository:** https://github.com/cvs-health/langfair
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.2.0
