### 2024-12-03 #7558 @crvernon
:wave: @dylanbouchard - I'll take this one on myself.  While I am getting you some reviewers, please take care of the following:

- [ ] Reduce the paper length to around 1000 words
- [ ] Please add in the missing DOI listed above in your bib file where applicable

Thanks!
