### 2024-12-03 #7558 @dylanbouchard
> @editorialbot assign me as editor
>
> 👋 @dylanbouchard - I'll take this one on myself. While I am getting you some reviewers, please take care of the following:
>
> * [ ]  Reduce the paper length to around 1000 words
> * [ ]  Please add in the missing DOI listed above in your bib file where applicable
>
> Thanks!

Hi @crvernon! I have reduced the word count by ~800 words and fixed the DOI in bib file where applicable. Please let me know if reducing the paper length further is required. Thank you very much!
