### 2024-12-03 #7558 @dylanbouchard
> **Submitting author:** @dylanbouchard ([Dylan Bouchard](http://orcid.org/0009-0004-9233-2324)) **Repository:** https://github.com/cvs-health/langfair **Branch with paper.md** (empty if default branch): joss_paper **Version:** v0.2.0 **Editor:** @crvernon **Reviewers:** Pending **Managing EiC:** Chris Vernon
>
> ## Status
> [![status](https://camo.githubusercontent.com/6cef53af033bba155b36baccc6d48f8ec56fcb2ab62e882b4094ce58d22c9807/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f30666263303964653766613465383733616336386333636233306166646436362f7374617475732e737667)](https://joss.theoj.org/papers/0fbc09de7fa4e873ac68c3cb30afdd66)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @dylanbouchard. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @dylanbouchard if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

For editors, it seems that the following individuals may be a good fit:
- Erik-Jan van Kesteren
- Guanqun Yang
- Madiha Mirza
