### 2024-12-04 #7558 @xavieryao
> 👋 @xavieryao - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Happy to help. I will start reviewing right away.
