### 2024-12-04 #7570 @crvernon
👋 @dylanbouchard, @xavieryao, and @emily-sexton - This is the review thread for the paper. All of our communications will happen here from now on.
