### 2024-12-21 #7570 @dylanbouchard
> @dylanbouchard - I can see that there's a reference, 'Beyond the imitation game...' that has the author listed as 'authors, B.' When I go to the reference, it looks like there are hundreds of authors. Is this a way of referencing a paper with too many authors to list? Also, were the missing DOIs mentioned about addressed? @crvernon

@emily-sexton, that is my understanding yes. All of the references are auto-populated using bibtex, so that would be my guess.
