### 2024-12-22 #7570 @dylanbouchard
@emily-sexton good question, I will look into this and see if I can find anything.

Note that `venv` is not required. Other alternatives to `venv` can work too, like `virtualenv` or `conda` environments.

If you use conda, you can try:
```conda create --name langfair python=3.9```
```conda activate langfair```
```pip install langfair```

 If not, it might be worth trying with `virtualenv`. It seems like it may be easier to specify the version of Python in some of these alternatives.
