### 2024-12-29 cvs-health/langfair#82 @xavieryao
Utility for checking FTU
```python
# Check for fairness through unawareness
attribute = 'race'
df = pd.DataFrame({'prompt': prompts})
df[attribute + '_words'] =  cdg.parse_texts(texts=prompts, attribute=attribute)

# Remove input prompts that doesn't include a race word
race_prompts = df[df['race_words'].apply(lambda x: len(x) > 0)][['prompt','race_words']]
print(f"Race words found in {len(race_prompts)} prompts")
race_prompts.tail(5)
```

**Describe the solution you'd like**
Wrap the above snippet (from the counterfactual example notebook) as an utility function (or class).

Add instructions / examples to the main document on how to check FTU.
