### 2024-12-29 #7570 @xavieryao
@dylanbouchard great work! This sets an example of high quality (open source) software project.

Some minor issues that can be easily patched:

- [ ] Minor text errors in the readme and a notebook: https://github.com/cvs-health/langfair/pull/81
- [ ] Omit the brackets when referring to the author as part of a sentence: “For a case study on writers block, see Upper (1974).” Such as at line 71 of the paper. See [JOSS guideline](https://joss.readthedocs.io/en/latest/paper.html#citations).
- [ ] Paper line 65: can you add an (informal) introduction of FTU? I understand that you have formal definitions in the technical companion, but it would be more helpful for the readers.
- [ ] Paper line 118: explicitly mention that the steps are what have been described previously?
- [ ] Adding the missing DOIs if possible.

My understanding is that the metrics for recommendation and classification are not specific to LLMs. I see that it is convenient to have them in one package, as these are also applications of LLMs. Could you add a discussion of existing repositories that offer those **generic** fairness metrics? Examples:
https://github.com/tensorflow/fairness-indicators
https://github.com/Trusted-AI/AIF360
https://github.com/amazon-science/generalized-fairness-metrics?tab=readme-ov-file

Is there a clearer pointer to "off-the-shelf FTU check" (line 67) other than in an example notebook?  https://github.com/cvs-health/langfair/issues/82
