### 2025-01-01 #7570 @xavieryao
@dylanbouchard Thank you for the quick responses. All of my concerns have been adequately addressed and I think both the paper and the project are of great quality.

@crvernon I have no further comments and would recommend **Accept**.

Happy new year.
