### 2025-01-02 #7570 @dylanbouchard
@crvernon two questions:
1. Should the submission be updated to the version that includes the updates recommended and approved by @xavieryao, i.e. v0.3.1?
2. Is there a preferred way to handle text contained in backticks from spilling over the margin? This appears to be happening in line 55 with the mention of `CounterfactualGenerator`
