### 2025-01-03 #7570 @crvernon
@dylanbouchard - yes to the version update
