### 2025-01-03 #7570 @crvernon
@dylanbouchard I also have an open request in for our development team to check out how to best represent the overflow generated by using backticks.  I'll let you know when I hear back.

Concerning the @emily-sexton raised about the many author paper and how to cite it:

This is the correct reference format to add to your bib file:

> @article{srivastava2022beyond,
  title={Beyond the imitation game: Quantifying and extrapolating the capabilities of language models},
  author={Srivastava, Aarohi and Rastogi, Abhinav and Rao, Abhishek and Shoeb, Abu Awal Md and Abid, Abubakar and Fisch, Adam and Brown, Adam R and Santoro, Adam and Gupta, Aditya and Garriga-Alonso, Adri{\`a} and others},
  journal={arXiv preprint arXiv:2206.04615},
  year={2022},
  doi={10.48550/arXiv.2206.04615}
}

Also, for your references containing no date:  please either cite a release of these repositories or simply use the weblink in text.  The citation is preferred.  Thanks!
