### 2025-01-04 #7570 @emily-sexton
> > @dylanbouchard - I'm having troubles installing langfair because I have Python 3.13.1 and looks like langfair can only be used with Python <3.12, >=3.9. I'm trying to install an old version of python in a virtual environment but the best documentation i can find on stack overflow is many years old (https://stackoverflow.com/questions/5506110/is-it-possible-to-install-another-version-of-python-to-virtualenv) and I'm struggling.
> > Can you give some guidance on how best to install langfair if I have a newer version of python? Should I be installing an older version of python in a virtual environment? If so, do you know the best way to do that? Thanks!!
>
> @emily-sexton I looked into this more and it I was able to specify the Python version with `venv` (e.g. 3.9) as follows:
>
> `python3.9 -m venv myenv`
>
> Please let me know if you have further questions. Thank you!

@dylanbouchard - Thank you very much! I was able to create a virtual environment with python 3.9 and install langfair properly.
