### 2025-01-04 cvs-health/langfair#93 @emily-sexton
Errors when importing StereotypeMetrics
When I try to import StereotypeMetrics:

from langfair.metrics.stereotype import StereotypeMetrics

I get a number of errors, which I've included in an attached text file. The code still finishes with exit code 0.
[StereotypeMetrics_Error.txt](https://github.com/user-attachments/files/18308106/StereotypeMetrics_Error.txt)

When I run the following code:
sm = StereotypeMetrics()
stereo_result = sm.evaluate(responses=responses, categories=["gender"])
print(stereo_result['metrics'])
It's still calculating the metrics as expected, printing them, and finishing with exit code 0, so I'm unsure as to why it's throwing so many errors.

I'm using a Mac OS 14.7.2 and running Python 3.9 using virtualenv
