### 2025-01-04 #7570 @emily-sexton
@dylanbouchard - I just finished reviewing and think your work here is great! I have two points to bring up:

- [ ] I'm getting errors when I import StereotypeMetrics, although it still seems to be running properly. I've created a bug for that here: https://github.com/cvs-health/langfair/issues/93
- [ ] When I use the code provided in the usage examples, anytime I run one of the lines with 'await', I get an error that the await keyword is outside of an async function. Putting the code inside an async function works better for me. Do you think this should be added to the documentation or is there a better way that you get the lines with 'await' to run?

Everything else looks great and once those two things get sorted, we should be good to go!
