### 2025-01-04 cvs-health/langfair#93 @dylanbouchard
Thanks for creating this issue @emily-sexton! Are you running the code in a notebook or a .py file? If it is a .py file, the below might solve it

```
def main():
    # Your code here

if __name__ == '__main__':
    main()
```
