### 2025-01-04 #7570 @dylanbouchard
Thank you very much @emily-sexton! I have responded to your issue with a likely solution. I believe including the below in your .py file should fix both issues you state above:

```
def main():
    # Your code here

if __name__ == '__main__':
    main()
```
