### 2025-01-05 cvs-health/langfair#93 @emily-sexton
Hi @dylanbouchard - I'm running in a .py file in pycharm. You're right - the above code you provided solved the problem. Thank you! I'll close the ticket.
