### 2025-01-05 #7570 @emily-sexton
@dylanbouchard - Thank you for your response! Both issues are fixed with the solution you provided!

@crvernon  - Review is complete on my end and all good to accept!
