### 2025-01-06 #7570 @crvernon
@dylanbouchard - no solution was found unfortunately.  In this case, manually hyphenating would be the easiest appropriate solution.  Thanks!
