### 2025-01-06 #7570 @dylanbouchard
@crvernon I have fixed the missing DOIs and the text spilling into the margin. I think we should be good to go!
