### 2025-01-08 #7570 @dylanbouchard
@crvernon do the references look ok now? Happy to make additional corrections if needed. Thank you!
