### 2025-01-09 #7570 @dylanbouchard
Hi @crvernon, thank you very much for following up. Please see the requested information below:

- I have created the [Zenodo archival deposit](https://zenodo.org/records/14622998). The DOI is: 10.5281/zenodo.14622998
- Version 0.3.1 of LangFair (consistent with this submission) has already been released. You can find the release [here](https://github.com/cvs-health/langfair/releases/tag/v0.3.1).
