### 2025-01-17 #7570 @crvernon
@dylanbouchard the version number in your archive doesn't match the version of your release.  Please edit the metadata in your Zenodo archive to reflect the version of the release in your GitHub repo. Thanks.
