### 2025-01-17 #7570 @dylanbouchard
@crvernon thank you for catching that! I have updated the version on the Zenodo archive.
