### 2025-01-23 #7570 @crvernon
@dylanbouchard the paper is looking great.  I just need you to handle one last issue before I accept this for publication:

Please be sure to maintain capitalization where necessary in your references.  You can do this in your bib file by wrapping characters that you wish to maintain formatting of in curly brackets.  The only one I really see is on LINE 164, where "north american" should be "North American".

Once you fix this, I'll move to accept.  Thanks!
