### 2025-01-23 #7570 @dylanbouchard
Thank you @crvernon! I have just made the the requested change to the references and checked that the compiled PDF displays as it should. Please let me know if anything further is needed from my end.
