### 2025-01-23 #7570 @dylanbouchard
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/12931692378?check_suite_focus=true). The generated XML metadata file is invalid.
>
> Element citation, attribute 'key': [facet 'maxLength'] The value 'Arshaan_Nazir_and_Thadaka_Kalyan_Chakravarthy_and_David_Amore_Cecchini_and_Thadaka_Kalyan_Chakravarthy_and_Rakshit_Khajuria_and_Prikshit_Sharma_and_Ali_Tarik_Mirik_and_Veysel_Kocaman_and_David_Talby_LangTest_A_comprehensive_2024' has a length of '228'; this exceeds the allowed maximum length of '128'.

@crvernon I should be able to just use a different key for this one, let me see
