### 2025-01-23 #7570 @dylanbouchard
@crvernon I just removed the duplicate instance of this reference in the bib file. Thank you.
