### 2025-01-23 #7570 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
